textures "working"
Textures work. The way the assets are being transferred isn't
ideal though. Currently they are all being bundled up into the
final binary with include_dir!. The png images take forever
to generate image buffers so the performance is bad. The cube
with a 1024x1024 takes minutes to import. Something different
will need to be done with gltfs, which I've known for a while.